DocumentsUI : Fixing FragmentManager crash in Monkey run.
UT
============================================================
PRECONDITION :- Fill the device storage to ~95%

VERIFICATION COMMAND :- adb shell monkey --throttle 1000 -s 54321
                        --pct-touch 0 --ignore-crashes --ignore-timeouts
                        --ignore-security-exceptions -p
                        com.android.documentsui -v -v -v 200000

RESULT :- "java.lang.IllegalStateException: FragmentManager is already executing transactions".The above crash in log is not seen.

REASON FOR THE CHANGE :- commitNow() will Commits the current transaction synchronously. calling commit() doesn't perform the transaction immediately. Rather, the transaction is scheduled to run on the main UI thread as soon as it is able to do so. commitNow() will not execute all previously committed transaction,  it will only commit that transaction. Hence, it is incompatible with addToBackStack.
============================================================

Change-Id: I946e1c4a24b5779746b74ac6703e8b1833fae7a2
Signed-off-by: Shubhamoy Roy Choudhury <quic_shubhamo@quicinc.com>