[DocsUI, Search]: Adds basic dropdown options.
When search becomes active the type chips are to be replaced with 3
dropdowns: location, recency and type. The location allows one to select
how wide (everywhere) or narrow (this folder) the search is applied. The
recency allows one to limit the results to files modified in the last n
days for n = 1, 2, 7, 30 and 365. Finally, the type allows one to limit
the results to 4 broad categories (documents, videos, images and audio).
This CL provides the basic building UI blocks, without coding all
necessary logic. This is left for the future changes.

Bug: 391232249
Test: m DocumentsUIGoogle
Test: m SearchOptionsControllerTest
Flag: com.android.documentsui.flags.use_search_v2_read_only
Change-Id: I4d9006d18cf65ef818b7ff7cf014534ff1e8eaac