[DocsUI M3] Restyle Sort by menu bottom sheet
* In order to show hover effect on the bottom sheet menu items,
we need to add clickable/focusable on the level item, which
unfortunately breaks the onClick and onKey event, we need to
explicitly handle them.
* The "checkMark" from the CheckedTextView component shows a weird
ripple effect, to avoid that we need to use compound drawable
manually to show the check mark.

Check the attached bug for the demo.

Bug: 404985320
Test: m DocumentsUIGoogle && manual inspection
Flag: com.android.documentsui.flags.use_material3
Change-Id: I67249c67b4351aea93141ed4c3c953bd15215940