Clear selection when a search chip is pressed
Currently when a search chip is pressed the selection in the action bar
is retained. This then gets into a funky state where it can't be
cleared. To avoid this happening, let's ensure the selection is cleared
when a search is invoked via the search chips.

Bug: 405288522
Test: atest com.android.documentsui.SearchViewUiTest
Flag: com.android.documentsui.flags.use_material3

Change-Id: Ibf95bbe5813222cba2ea65c4ce46b29a2d0f226d