[DocsUI, Search]: Enable listing of archives.
Adds a code that detects if the listed document is an archive. If so, it
calls local archive provider to create a loader that allows us to access
the archive content. Saves the ContentProviderClient in document result,
so that it can be closed when document result is closed.

Bug: 385806471
Test: m DocumentsUIGoogle
Flag: com.android.documentsui.flags.use_search_v2_read_only
Change-Id: Ie35ab4bf9a5a04a4223caf3b28a5acf8151932e6