Only hide the AppsRow on devices the are non-FEATURE_PC
As part of the Material 3 uplift, devices which have more screen estate
should hide the AppsRow as the same information is available in the
RootsFragment. This updates the tests to skip when FEATURE_PC is
defined and the use_material3 flag is enabled.

Fix: 406344785
Bug: 383932124
Test: atest com.android.documentsui.dirlist.AppsRowManagerTest
Flag: com.android.documentsui.flags.use_material3

Change-Id: I839bfddc1a4b35a7e7fb67cc1b9c7f7d8a0c1a9f