Fix up error on testPasteIntoFolderOnRoot
The test SidebarUiTest#testPasteIntoFolderOnRoot is currently failing
as the view that appears needs to cover 90% of the screen. As the view
is the right click menu this is not true. This appears to be a bug so
let's create our own click action to remove the constraint and use that
to perform the action instead.

Fix: 406361706
Test: atest com.android.documentsui.SidebarUiTest
Flag: EXEMPT bug fix
Change-Id: Idf15f3fa7de4207e4b0401200bd2bbf13f0a4f64