Show selection mode when search bar is available on Recents
The search bar currently takes over the entire top of DocumentsUI when
enabled. This worked well when the selection bar was an ActionBar as it
simply overlaid the entire header. Since we've moved to the navigation
view, it no longer works. This adds a test to show the selection when
activated and hide the search bar.

Bug: 405807817
Flag: com.android.documentsui.flags.use_material3
Test: atest com.android.documentsui.FilesActivityUiTest
Change-Id: Ib3f346883fb052a27609f79d2bc1465a9bec7105