Don't browse nested archives
Restrict the visibility of the "Browse", "Extract here", "Open" and
"Open With" menu items.
The "Browse", "Open" and "Open with" menu items are not visible for
nested archives (i.e. archives located in a browsed archive).
The "Extract here" menu item is not visible if the current directory
does not allow the creation of subdirectories.

Bug: 406872038
Flag: com.android.documentsui.flags.use_material3
Flag: com.android.documentsui.flags.zip_ng_ro
Test: atest DocumentsUIGoogleTests:com.android.documentsui.files.MenuManagerTest
Test: atest DocumentsUIGoogleTests:com.android.documentsui.picker.MenuManagerTest
Test: atest DocumentsUIGoogleTests:com.android.documentsui.base.DocumentInfoTest
Change-Id: I717c634c9de34d0ecfd05eec544cc30186d4ffa5