Introduce new TestFilesRule to extract file creation
This converts the BandSelectionUiTest to JUnit 4 and introduces a new
Rule that enables creation of test files to remove the odd nature of
file creation by overriding the method. It also migrates all the
default usages of initTestFiles to the Rule.

In future CLs it will introduce a Builder pattern to customise the base
files that will be created.

Bug: 407646510
Test: atest com.android.documentsui.BandSelectionUiTest
Flag: EXEMPT test change only
Change-Id: I81166f2112da2ab6e6f76be7143c239321bae2c1