Create builder pattern to initialise custom test files
Most tests simply rely on the default folder and file creation, however
some of the test use custom logic. This introduces the ability to
create custom directory structures using the TestFilesRule.

Along with this it removes the mDocsHelper logic from the
ActivityTestJunit4 class. The resetting of storage removes the files
on disk but doesn't clear them from the database, this leads to odd
test behaviours where you can see the file but not perform actions on
them.

Bug: 407646510
Test: atest com.android.documentsui.FilesActivityUiTest
Flag: EXEMPT test change
Change-Id: Ic111d66dab423dec7edabb60baf45bdbb06b9fd9