Always include intent-filters for PickActivity
The PickActivity is currently listed as priority 100 for all the
intent-filters. These will always be smaller than TrampolineActivity
when enabled as well as Photopicker so we don't really have to remove
them. It adds the safeguard that when the flag is off it falls back.

Given for this flag we'll just be launching it to handle the intents
related to Photopicker. Let's just include the GET_CONTENT in the
TrampolineActivity and we can include others when the need arises.

Bug: 377771195
Test: atest com.android.documentsui.TrampolineActivity
Flag: com.android.documentsui.flags.redirect_get_content_ro
Change-Id: I4abe21989dcc20b20e98b0d28c8aac312090c73d