[DocsUI M3] Add selection icons for grid items
Add a selection circle that when tapped selects the item and displays
a check mark. The selection circle is now the selection region.

This also preserves the logic of not having a selection region for
folders in picker mode.

If the user selects the item another way, the check will be shown as
well.

The touch target for the selection circle is 34x34 but the visible
circle is smaller, defined by selection_circle.xml.

See bug for demo.

Bug: 407853191
Test: m DocumentsUIGoogle && manual inspection
Flag: com.android.documentsui.flags.use_material3
Change-Id: Ib44fd8a0c5780206c1d3cc11ea2799f91f0eebb9