Fix bugs when cancelling jobs
When cancelling the current job, the service would try get the in
progress notification of the next item, which would be in the
STATE_CREATED (queued) state, and that would cause a crash.

Additionally, cancelling a job that hadn't started yet would never
inform the service that the job had finished, so it would never be
removed from the list of tracked jobs.

Bug: 409489330
Bug: 400352714
Test: atest -c 'DocumentsUIGoogleTests:com.android.documentsui.services'
      and manual
Flag: EXEMPT bugfix
Change-Id: I114f684548afdc1328e0e3fee1f3cc03ffe53a81