Fix crash when opening overflow menu whilst selection active
When the material 3 flag is enabled, the NavigationViewManager takes
care of updating the options menu. This means we don't delegate it to
the menu manager directly but we do it through the NVM. Previously a
call to the MenuManager was missed in the Picker which caused a null
dereference. Update that call site and add a test to avoid a regression

Fix: 408935808
Test: atest PickActivityTest#testOptionMenuWorksWhileOptionSelected
Flag: com.android.documentsui.flags.use_material3
Change-Id: I5a692c7593327c0493bbe38bdd4a1289f0761cb6