[DocsUI] Fix functional tests failures related to PeekBot dependency
The PeekBot is initialized with a matcher that makes use of the
MaterialToolbar class. It throws a java.lang.ClassNotFoundException
when M3 is not enabled.

To avoid similar issues in the future, initialize the PeekBot directly
in tests for the Peek feature. The PeekBot is only initialized when Peek
is enabled.

Bug: 408956866
Flag: TEST_ONLY
Flag: com.android.documentsui.flags.use_material3
Flag: com.android.documentsui.flags.use_peek_preview_ro
Test: atest DocumentsUIGoogleTests:com.android.documentsui (with M3
enabled and disabled)
Change-Id: Ie1fcfa8548c9a190894e1c8fd0bc66bbccc69f2f

Change-Id: I745c0182dd2bc58a62069d5c9a4570922469fd60