Mark bots in ActivityTestJunit4 as lateinit var
When tests attempt to use bots, as it's current a nullable property
they have to use the non-null assertion everywhere. Given the @Before
in ActivityTestJunit4 will run prior to the test cases, it's safe to
use lateinit var for bots and remove all the non-null assertions in
tests.

Bug: 407646510
Test: atest DocumentsUIGoogleTests
Flag: EXEMPT test change
Change-Id: Iae229c78a354b0ebc732cc8cadb0126379dbd906