Update the top bar color for all layouts
The status bar for all layouts was a slightly different color. It looks
like the background was actually being inherited from the coordinator
layout underneath it whilst the linearlayout that houses the roots and
main container started with a margin off the top. Simply setting the
background color for the coordinator layout ensures the status bar
color is set.

Fix: 408107825
Test: m DocumentsUIGoogle and manual inspection
Flag: com.android.documentsui.flags.use_material3
Change-Id: Ia0217066fe1b68c284c34b0e060f21bf50e28f48