Log the initial screen off and sleep timeout values
There appears to be an emerging cluster of errors that seem to find both
the initialScreenOffTimeoutValue and initialSleepTimeoutValue values as
null. This is puzzling as the values should be retrieved in the setUp
then restored in the tearDown. Let's log these values as warnings to we
can inspect logcat to see if they're being retrieved during setup. My
initial hunch is that either the value being returned is null OR the
setup is failing and the teardown is always being run. This should shed
some more light though.

This also removes the unused resetStorage which stopped being used in
http://ag/32833099.

Bug: 409044744
Bug: 408942389
Test: m DocumentsUIGoogle
Flag: EXEMPT test only change
Change-Id: I3772544b8a62bf5621b0034cd08d3a4cff9d8ae8