Trim the application name to make it safe for presentation
The application name is presented in the ConfirmFragment and as such we
don't want to allow for any length. This follows a similar approach to
PackageManager using the TextUtils.makeSafeForPresentation with a total
available character length of 500.

This removes the unused getCallingAppName from the DirectoryFragment as
it was causing false positives from DirectoryFragment to avoid false
positives when trying to find who calls the Shared function.

On top of this, add some quotation marks around the app name to avoid
the app name being a contination of the existing text in the dialog,
e.g. 'This will let app name access current and future content storage
in Alarms' will now be 'This will let "app name" access current and
future content storage'.

Bug: 397216537
Test: atest com.android.documentsui.picker.ApplicationNameTest
Flag: EXEMPT bug fix
Change-Id: Iad0d03de09b1e4ad953bd6bd46a619cfcc56d384