[DocsUI M3] Ignore live scaling in the column count calculation
Ignore the live scaling factor in the column count calculation
as the grid items are a fixed size and are not affected by pinch
zoom (the action that triggers the live scaling factor to change).

This means that when the user pinch zooms, there will be no change.
Previously the amount of white space would increase and decrease which
looked strange.

See bug for demo.

Bug: 407851636
Test: m DocumentsUIGoogle && manual inspection
Flag: com.android.documentsui.flags.use_material3
Change-Id: If969b3cdb46700954588e589e7940a2306f041b3