Ensure createDocument has a valid RootInfo to create it at
This is currently failing with the following error:

  java.lang.IllegalArgumentException: Can't create file file_1 in null
  parent.

This is because the root used is incorrect, use the appropriate root.

Unfortunately some of the tests also fail when the use_material3 flag is
enabled, so this updates the SortBot to check not only if the
table_header is visible but also if the column to sort by is visible.
This is because on use_material3 the Name table header is shown and
clickable, which causes a false positive.

Fix: 408352609
Test: atest com.android.documentsui.SortDocumentUiTest
Flag: EXEMPT test change
Change-Id: I87c2903849b461a1061ff02e622938e4e6de0017