[DocsUI M3] Make grid item height flexible
Previously the grid item height was fixed but this became
an issue when changing the font or display size via Settings.
The text in the nameplate can become too large for the height
of the nameplate leading it to be cut off. Instead make the height
of the nameplate flexible so it wraps the text. Also do the same for
the entire grid item so that the nameplate can flex.

Every grid item will have the same height even if one of its two
lines text in the nameplate is not present as the GridLayoutManager
gives all items the same height.

See bug for demo.

Bug: 399460020
Test: m DocumentsUIGoogle && manual inspection
Flag: com.android.documentsui.flags.use_material3
Change-Id: If012e80b2045c86027ea8cc7dc2e611315dfdd9c