[DocsUI M3] Uplift disabled state for grid
Apply a 60% opacity to the whole item, not just the
thumbnail.

Also ensure that when the thumbnail gets loaded it is shown
with 100% opacity (from the initial 0% opacity). Previously
this was taken care of during the enable/disable logic applied
to the thumbnail but this is no longer the case.

See bug for screenshots.

Bug: 409878953
Test: m DocumentsUIGoogle && manual inspection
Flag: com.android.documentsui.flags.use_material3
Change-Id: I9017b19ad067ee53fd887eb8fd92e7b78dcd6474