[DocsUI M3] Hide the second row in the nameplate when there's no text
When there's no date, instead of setting the date text to null,
just hide the date TextView. This is how the details TextView
was already handled.

If there is no date and not details, then the entire second row
of text will be properly hidden and the nameplate will just wrap
the title (a feature added in ag/33002418). Previously, an empty
gap would be left for the second row which looked strange.

See bug for screenshots.

Bug: 409907555
Test: m DocumentsUIGoogle && manual inspection
Flag: com.android.documentsui.flags.use_material3
Change-Id: Ib5488626f3b4ef82255ab1ba4864911ee64d0c68