[DocsUI M3] Update root icons
This is a preparation to remove icon overrides in desktop RRO.

Previously these icons are overwritten in the RRO, in order to
remove the unnecessary RRO icon overrides, this CL updates the
icons to the latest ones from the Material icons.

Check the attached bug for demo.

Bug: 405236942
Test: m DocumentsUIGoogle && manual inspection
Flag: com.android.documentsui.flags.use_material3
Change-Id: I6ecc68cb176896c9d96c9083a53e8d8ed6934345