[DocsUI M3] Rename PickFragment to PickDirectoryFragment
The OPEN_TREE/PICK_COPY_DESTINATION pickers use an additional
SaveFragment/PickFragment on top of the PickActivity to add the
Cancel and Pick buttons for users to pick a folder. Rename the
PickFragment to PickDirectoryFragment to make this clearer.

See bug for demo.

Bug: 409931248
Test: m DocumentsUIGoogle && manual inspection
Flag: com.android.documentsui.flags.use_material3
Change-Id: I2794cf5d64195c50cfb8374aab20799c2ab38b8c