Ensure the options menu is refreshed on Recent when no selection
When selecting an item the app bar menu is refreshed to show actions
relating to the current selection. When the selection is cleared, the
menu should refresh. Unfortunately one of the sections was incorrectly
bailing out when the search bar was showing, this refactors some of the
update() method to be clearer and ensures the action menu is always
updated.

This also includes a fix to the openDrawer method which was not
correctly identifying the drawer when in the nav_rail layout. The
selector was finding `container_roots` when the layout actually uses
`nav_rail_container_roots`. It updates it to use regex instead of exact
match.

Fix: 411295692
Test: atest com.android.documentsui.FilesActivityUiTest
Flag: com.android.documentsui.flags.use_material3
Change-Id: Ic8c5dc768ca5135392749c91bfbfdf684e39a4e6