Stub out activity.registerReceiver in files.MenuManagerTest
MenuManager initiates the JobPanelController on creation, and the latter
calls registerReceiver on the passed in activity on init so it can
receive broadcasts. However the test activity created in MenuManagerTest
does not handle this properly, so stub it out as it is not needed for
the test.

Bug: 412583130
Test: atest -c 'DocumentsUIGoogleTests:com.android.documentsui.files.MenuManagerTest'
Flag: com.android.documentsui.flags.visual_signals_ro
Change-Id: I2d98f4fcc853adf9e4a477824001f3cbef6cb85a