Fix BandSelectionUiTest when use_material3 flag is enabled
When the use_material3 flag is enabled, the size of the items in the
grid mode is different. This can cause inconsistencies in the number of
items that are selected. The test is actually fundamentally flawed as
when it is ran on a larger screen the number of items differs as well.
This updates the test to dynamically calculate the number of items in
the top row and assert the selection is based on that result.

Bug: 412776499
Test: atest com.android.documentsui.BandSelectionUiTest
Flag: com.android.documentsui.flags.use_material3
Change-Id: Ia987461d9b895b5c5f9d3e6bec3dd7c797075864