Fix up findSelectionHotspot logic for use_material3
The current logic for the findSelectionHotspot (which underpins the
logic to selectDocument) doesn't work for grid view. It attempts to find
the `icon_wrapper` ID in the view hierarchy. When in grid mode and with
the use_material3 flag enabled the selectable region is actually
selection_circle instead. Update the logic to respect that.

Bug: 412776499
Test: atest com.android.documentsui.CancelFromNotificationUiTest
Flag: com.android.documentsui.flags.use_material3
Change-Id: If9f997455108a38f503b8f22542d166c40ccdac2