DocsUI, Search]: Search dropdown notifications.
Adds a search options listener. Adds a class that remembers a snapshot
of the options. Adds a method for setting a option change listener. Adds
methods for recording state of options in a bundle, to make it
compatible with the SearchChipViewManager approach. Triggers a new
search request if the search options change.

Bug: 391232249
Test: m DocumentsUIGoogle
Test: m SearchOptionsControllerTest
Flag: com.android.documentsui.flags.use_search_v2_read_only
Change-Id: I727eae423c14d7ef9c5a3c222cbff10e3e7ce868