Simply check if the nav bar is visible before adding padding
It turns out all the excessive API calls to verify that the nav bar is
visible is not necessary as there appears to be a simple API that
returns true/false if the gesture bar is visible or not. Let's prefer
that instead.

Bug: 402604097
Test: m DocumentsUIGoogle to Pixel tablet and observe
Flag: com.android.documentsui.flags.use_material3
Change-Id: I318c5062502c1001f19ac85ceba87be1fe1d34ed