Add test to assert selection bar shows up when search item is selected
As part of the material 3 refresh, the selection bar was rewritten to
not use an ActionBar. There are a number of implicit behaviours bulit
into the action bar that need to be hoisted into the new code.
Unfortunately it looks like these aren't covered by tests.

This test asserts that when a search is currently happening, selection
of any items in the search view should show a "N selected" bar up the
top.

Bug: 412895530
Test: atest com.android.documentsui.SearchViewUiTest
Flag: EXEMPT adding test to validate behaviour with flag off
Change-Id: Ibcd303d804514dda1ffb8a1455a63ac8305a74bf