[DocsUI M3] Fix GridEvenSpacingDecorationTest by setting measured width
Follow up to ag/32962378 that changed from using parent.getWidth() to
parent.getMeasuredWidth() in GridEvenSpacingDecoration. The
GridEvenSpacingDecorationTest was not updated to match this. It was
using setLeft() and setRight() which works to set the width but not the
measured width. Use setMeasuredDimension() instead.

Bug: 412919167
Test: atest GridEvenSpacingDecorationTest
Flag: com.android.documentsui.flags.use_material3
Change-Id: I519e34de5f8df51afba8d50c1e78d4d6e609973b