Handle queued state for progress messages
Currently, we delay file URI resolution to when the operation actually
starts as resolution could require I/O in cases where there are many
files in an operation (in which case the list of URIs are passed through
a temporary file). However, we need the filename for single file
operation progress messages while the operation is still queued, so in
that case only, resolve the file URI early.

This also refactors duplicated code in jobs to the resolved resources
common job.

Bug: 407674989
Test: atest -c 'DocumentsUIGoogleTests:com.android.documentsui.services'
Flag: com.android.documentsui.flags.visual_signals_ro
Change-Id: Ic6bc290e48554f29d0ad418a3bac4f9b25f00b10