Update tests to wait and assert vs. immediately asserting
An upcoming update will change the way the selection bar is setup.
Currently the selection bar deflates and inflates a different menu in
the same toolbar. This means the assertions are performed after the
inflation. In an upcoming update this update will happen separately and
thus the assertions turn flaky. This has no functional change and still
has the same assertions, just more robust.

Bug: 412895530
Test: atest com.android.documentsui.FilesActivityUiTest
Test: atest com.android.documentsui.SearchViewUiTest
Flag: EXEMPT updating test to be more robust for upcoming changes
Change-Id: I49d236ca8379f23bafed0ed97ea9a77b32d9e2f6