Update selection logic to use a separate MaterialToolbar
The selection bar was updated (behind the material3 flag) to update the
existing MaterialToolbar with the action mode menu. Unfortunately that
doesn't seem to play nice with the ActionView / ActionProvider pattern
(which the SearchView is using).

It seems the only way to achieve a detached selection bar (i.e. not an
actionmode bar) is to have a hidden secondary MaterialToolbar which
shows / hides when a selection is made.

This also largely walks back the changes that were made in ag/31787743
as well given they are not required anymore.

Bug: 412895530
Test: atest com.android.documentsui.FilesActivityUiTest
Test: atest com.android.documentsui.SearchViewUiTest
Flag: com.android.documentsui.flags.use_material3
Change-Id: I28877a84bd75a6ca584c125feeaeb38ac7dd81f2