[DocsUI M3] Ensure no race condition when clearing selection in tests
Previously selectDocument() was used to deselect files (as
clicking on the selection region of a selected document deselects
it). Unfortunately this function doesn't wait for the deselection
to complete which can lead to test failures. Instead introduce
clearSelection() which clicks the cancel button and then waits for
it to disappear, indicating selection was cleared.

Bug: 411295692
Test: atest FileCopyUiTest
Test: atest FilesActivityUiTest
Test: atest InternalStorageUiTest
Test: atest PeekUiTest
Test: atest SearchViewUiTest
Flag: com.android.documentsui.flags.use_material3
Change-Id: Ieaf8464334af1cccc61d65b04b723ae84c018c3e