Don't close mActivityScenario when null
A previous CL overrode the `launchActivity` method, unfortunately due to
the sequencing of Junit, the @Before method of setUp in
ActivityTestJunit4 is ran which calls the `launchActivity` method. When
the overrode method is called it fails the assumption and bails early,
this didn't end up assigning anything to the mActivityScenario member
variable. The tearDown is always ran and as a result the
mActivityScenario!!.close() method was failing with a null reference
error.

Bug: 410436015
Test: atest DocumentsUIGoogleTests:com.android.documentsui.NavRailUiTest
Flag: com.android.documentsui.flags.use_material3
Change-Id: I1efd7a3bd57e0f69e651a4f9b2d6956a958c3b80