Fix InflateMessageDocumentHolderTest for use_material3
When use_material3 is enabled, the MessageDocumentHolder layout contains
com.google.android.material.progressindicator.CircularProgressIndicator
which attempts to perform an animation. The current way the test is
structured causes the inflation of the layout to happen on the main
thread which has no Looper thread avaiable. To Enable this animation to
work, move the initialization to the test and annotate the method
appropriately so it's ran on the UI thread and it's associated Looper
thread.

Bug: 412776499
Test: atest InflateMessageDocumentHolderTest
Flag: com.android.documentsui.flags.use_material3
Change-Id: I937dff745419f2a55284dcdaf3e6bf8e461fa0b6