[DocsUI, Search]: Show progress while searching.
Adds a new method onSearchStarting() to the SearchManagerListener. Uses
it to inform the search manager listener created in the base activity to
indicate that the model is loading. This triggers a progress bar in the
file list view (RecyclerView). Adds ability to explicitly set th model
state to loading state. Adjusts the tests.

Bug: 412504830
Test: m DocumentsUIGoogle
Test: m SearchViewManagerTest
Flag: com.android.documentsui.flags.use_search_v2_read_only
Change-Id: I76778412d9fbce1ff9039e9add959743762e1bf8