[DocsUI M3] Fix talkback issue on the breadcrumb
Previously we put clickable/focusable directly on the TextView
inside the breadcrumb, which doesn't work well with the special
handling (AccessibilityEventRouter) of a11y inside breadcrumb,
which causes double tapping can't activate the breadcrumb item.
This CL moves the clickable/focusable back to the breadcrumb item
container and use duplicateParentState=true on the TextView to
get the expected state on the TextView.

Check the attached bug for the demo.

Bug: 415695308
Test: m DocumentsUIGoogle && manual inspection
Flag: com.android.documentsui.flags.use_material3
Change-Id: I3d9349fef193e86ca15ba32be0b6ed39963e460f