[DocsUI] Update clearSelection() method to use "Cancel" text too
Follow up to ag/33304970 and ag/33406606 which only sets up the test
method clearSelection() correctly for when the useMaterial3 flag is
on. When this flag is on, the clear selection button inside the
selection_bar (ag/33285246) and has the content description as
"Clear selection" (ag/33384387).

However when the flag is off, the clear selection button is in the
action_mode_bar and has content description "Cancel". Update
clearSelection() accordingly.

Bug: 408898634
Bug: 416451224
Test: atest FileCopyUiTest
Test: atest FilesActivityUiTest
Test: atest InternalStorageUiTest
Test: atest PeekUiTest
Test: atest SearchViewUiTest
Test: atest PickActivityTest
Flag: EXEMPT fixing test failures
Change-Id: Ic17b453db3df78f6c814b06389480e122a53c78d