[DocsUI M3] Only show apps row when show_apps_row config is enabled
Follow up to ag/32733002 which only showed the apps row for
non-FEATURE_PC devices. Instead change the condition to only
show the apps row when the newly introduced show_apps_row config
is enabled. This is true by default but will be overwritten in a
desktop DocumentsUIGoogleOverlay to be false so that the apps row
is hidden on desktops.

Bug: 414692965
Test: atest com.android.documentsui.dirlist.AppsRowManagerTest
Flag: com.android.documentsui.flags.use_material3
Change-Id: I1dc85c376ed02540028941cef8e413a139a6c947