Don't run IconHelperTest on devcies with HSUM
On devices with HSUM the system user is not a real user. This means the
main user will typically NOT be the system user. Thus the test
testShouldShowBadge_returnFalse_onSystemUser will fail on these devices.
However, the system user doesn't ever use DocumentsUI on these devices
(as it's truly the system user) and thus the test is actually not
helpful on these devices.

Fix: 383281556
Test: atest com.android.documentsui.dirlist.IconHelperTest
Flag: EXEMPT Test fix
Change-Id: I4955e13c2b17a31ea05813e2e235501d5580b2c8