[DocsUI M3] Fix video root layout in navigation rail
Video root supports cross profile but it doesn't have a counterpart
root for work profile user, so a stub root item will be created in
this case, for Navigation rail layout, the stub item needs to be
created with the NavRailRootItem instead of the RootItem.

Check the attached bug for the before/after comparison.

Bug: 409929088
Test: m DocumentsUIGoogle && manual inspection
Test: atest DocumentsUIUnitTests:com.android.documentsui.sidebar.RootItemListBuilderTest
Flag: com.android.documentsui.flags.use_material3
Change-Id: Ia5b5cf0f3c0bedb362edbc7d9657f137520d054f