[DocsUI M3] Fix the top gap when force_material3 is OFF
When use_material3 flag is ON but force_material3 config is OFF,
the values-w900dp-600dp/config.xml in the falg on resource folder
still takes effect, the scrolling_behavior defined there causes
a big gap for certain screen sizes. This CL changes the file to be
config_m3.xml to make it explicit that it should only be used in
M3, and also update the config variable names to have m3 suffix.

Check the attached bug for the before/after comparison.

Bug: 418124756
Test: m DocumentsUIGoogle && manual inspection
Flag: com.android.documentsui.flags.use_material3
Change-Id: Ie29d7002d361cd66aa4c5631b28efe17b4162ceb