Ensure we type text into the search box instead of setting it
A previous change to make the search tests more robust ended up using
setText instead of typeText. This has the unfortunate side effect of not
focusing on the search text and thus failing the test when trying to
assert that search icon is visible on slower devices as it closes the
app. Ensure the view wait code is there but also move back to typing in
the text box instead.

This also inlines the SEARCH_INPUT variable as it was only used in one
place and with the addition of the docked search bar it would require
two of them and thus makes no sense to keep around.

Fix: 418404084
Flag: EXEMPT test fix
Test: atest com.android.documentsui.SearchViewUiTest
Change-Id: I85080149f68eaf015e075d8e477e3145c7f0f8f6