Update MenuManager tests to assert invisible select action icons
The MenuManager tests are currently failing as there are 2 tests which
assert that the "SELECT" button is enabled and visible. This is not true
when material 3 flag is enabled as they are moved to buttons at the
bottom of the screen. Update the tests appropriately.

Bug: 412776499
Test: atest com.android.documentsui.picker.MenuManagerTest
Flag: com.android.documentsui.flags.use_material3
Change-Id: Ie46ad6503fe20f54aca98f27f8e8cd53057856f7