Move class initialization to setUp method
The class initialization (which is always ran) was looking for
R.id.job_progress_indicator. Unfortunately this isn't available when the
USE_VISUAL_SIGNALS_RO flag is off (the ID is eliminated from the R
class). Move the initialization to the setUp method instead to ensure it
isn't ran when the flag is disabled.

Fix: 420543245
Test: atest com.android.documentsui.JobPanelControllerTest
Flag: com.android.documentsui.flags.visual_signals_ro
Change-Id: I08ea4d75e18f73b34f21e7136c2886cc653555fc