[DocsUI Peek] Metadata sheet container
Only the SideSheetBehavior is used from the side sheet material
component (dragging/expand/collapse animation and resizing of
sibling). This behavior is applied to the container of the metadata
sheet, which is transparent. A custom background with rounded corners
is applied for the metadata sheet itself. This is because a margin
issue raised in b/419417692, which prevents us from using the material
component as-is.

This implementation is only valid on large screens. The full
responsive layout will be implemented in later changes.

The metadatat sheet is systematically expanded (in a task posted via
`metadataContainer.post` after the fragment is created and visible,
because of b/419446581). Expand/collapse handling will also be added
in later changes.

Bug: 382163595
Bug: 419446581
Bug: 419417692
Flag: com.android.documentsui.flags.use_material3
Flag: com.android.documentsui.flags.use_peek_preview_ro
Test: PeekUiTest#testMetadataSheet
Change-Id: Ib0786faa52813b7cd96e0415d3bd7d2575f31ea0